pyproject.toml
[project]
name = "fastmarketing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.18",
    "browser-use>=0.1.41",
    "certifi>=2025.1.31",
    "pandas>=2.2.3",
    "playwright==1.51.0",
    "pydantic==2.10.6",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.40",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "ruff>=0.11.6",
]
